--- a/src/Humanizer/ArticlePrefixSort.cs
+++ b/src/Humanizer/ArticlePrefixSort.cs
@@ -91,7 +91,7 @@
 
         static string ToOriginalFormat(string append, string suffix)
         {
-            var suffixIndex = append.IndexOf(suffix, StringComparison.CurrentCulture);
+            var suffixIndex = append.AsSpan().IndexOf(suffix.AsSpan(), StringComparison.CurrentCulture);
             var insertion = append.Remove(suffixIndex);
             suffix = suffix.TrimStart();
             insertion = insertion.TrimEnd();