v2.14.1

| Method               | Mean     | Error   | StdDev  | Allocated |
|--------------------- |---------:|--------:|--------:|----------:|
| PrependArticleSuffix | 426.9 ns | 7.07 ns | 6.61 ns |   1.16 KB |

v3.0.0

| Method               | Mean     | Error   | StdDev  | Allocated |
|--------------------- |---------:|--------:|--------:|----------:|
| PrependArticleSuffix | 152.8 ns | 1.26 ns | 1.18 ns |     312 B |